+++ api/src/main/java/com/discoodle/api/repository/UserRepository.java
@@ -15,7 +15,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findUserByUserName(String username);
 
     @Query("SELECT user FROM User user where user.id = ?1")
-    Optional<User> findUserByID(Integer user_ID);
+    Optional<User> findUserByID(Long user_ID);
 
     @Query("SELECT user FROM User user where user.mail = ?1")
     Optional<User> findUserByMail(String mail);
